requirements.txt
beautifulsoup4
selenium
webdriver-manager
pandas
mysql-connector-python
